pyproject.toml
[tool.poetry]
name = "window-detector"
version = "0.1.0"
description = "A Python API that detects if a window is open in an image."
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
flask = "^2.0.2"
numpy = "^1.21.2"
opencv-python = "^4.5.3"
opencv-python-headless = "^4.5.3"
tensorflow = "^2.16.1"
tensorflow-metal = { version = "^1.1.0", markers = "sys_platform == 'darwin'" }
pillow = "^10.3.0"
scipy = "^1.13.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry.scripts]
ml-window-detector = 'window_detector.main:main'
